requirements.txt
CairoSVG==2.7.1
librosa==0.10.2.post1
Pillow==11.0.0
pygame==2.6.1
pyserial==3.5
